state_medical_boards/requirements.txt
dspy==2.5.29
matplotlib==3.8.4
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
scikit_learn==1.4.2
seaborn==0.13.2
tiktoken==0.8.0
tqdm==4.67.1
